requirements.txt
flask==3.0.2
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
gunicorn==21.2.0
requests==2.31.0 